Minor refactoring :)
