+++ _install/langs/en/install.lang.php
@@ -22,7 +22,7 @@ $LANG = array(
     'choose_install_lang' => 'Please choose your language to begin the installation',
     'requirements_desc' => 'WARNING! Please make sure <abbr title="Your Server. On local host, it will be you (your computer)">you are</abbr> connected to the Internet and your server has the <a href="' . Controller::SOFTWARE_REQUIREMENTS_URL . '" target="_blank" rel="noopener">minimum requirements</a>.',
     'requirements2_desc' => 'Before to continue, please create a MySQL database and assign a user to it with full privileges. Once you created the MySQL database and its user, make sure to write down the database name, username and password since you will need them for the installation.',
-    'config_path' => '&quot;protected&quot; directory path',
+    'config_path' => '&quot;protected&quot; folder path',
     'desc_config_path' => 'Please specify the full path of your &quot;protected&quot; folder.<br />
         It is wise and advisable (but not mandatory in any case) to put this directory outside of the public directory of the Web server.',
     'need_frame' => 'You must use a Web browser that supports inline frames!',
